% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt2samples.R
\docType{data}
\name{conditional}
\alias{conditional}
\title{Conditional with default settings}
\format{A data frame with 1000 rows and 4 variables:
\describe{
\item{\code{file}}{character - file identifier}
\item{\code{id}}{integer - sample and completion number as referenced in the openai gpt2 sample file}
\item{\code{text}}{character - the text, either a sample context or a completion}
\item{\code{type}}{character - indicator of whether the text is a sample context or a completion}
}}
\source{
\url{https://github.com/openai/gpt-2/tree/master/gpt-2-samples}
}
\usage{
conditional
}
\description{
Conditionally generated texts with contexts from WebText's test set
}
\keyword{datasets}
