% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt2samples.R
\docType{data}
\name{conditional_t07}
\alias{conditional_t07}
\title{Conditional with temperature 0.7}
\format{A tidy data frame with 24081 rows and 4 variables:
\describe{
\item{\code{file}}{character - file reference}
\item{\code{id}}{integer - sample and completion id number as referenced in the openai gpt2 sample file}
\item{\code{type}}{character - indicator of whether the text is a sample context or a completion}
\item{\code{text}}{character - the text, either a sample context or a completion}
}}
\source{
\url{https://github.com/openai/gpt-2/tree/master/gpt-2-samples}
}
\usage{
conditional_t07
}
\description{
Conditionally generated texts with contexts from WebText's test set
}
\keyword{datasets}
