% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt2samples.R
\docType{package}
\name{gpt2samples}
\alias{gpt2samples}
\alias{gpt2samples-package}
\title{GPT-2 Samples}
\description{
This package contains all sample data as generated by OpenAI's GPT-2 language model.
The data includes conditional texts with sample prompts and their completions, as well as
the unconditional texts with samples of generated text.
}
