% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt2samples.R
\docType{data}
\name{unconditional}
\alias{unconditional}
\title{Unconditional with default settings}
\format{A data frame with 19468 rows and 3 variables:
\describe{
\item{\code{file}}{character file identifier}
\item{\code{id}}{integer sample id number as referenced in the openai gpt2 sample file}
\item{\code{text}}{character - the text}
}}
\source{
\url{https://github.com/openai/gpt-2/tree/master/gpt-2-samples}
}
\usage{
unconditional
}
\description{
Unconditionally generated texts from Open AI GPT-2
}
\keyword{datasets}
