% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt2samples.R
\docType{data}
\name{unconditional_t07}
\alias{unconditional_t07}
\title{Unconditional samples with temperature 0.7}
\format{A tidy data frame with 28841 rows and 3 variables:
\describe{
\item{\code{file}}{character file identifier}
\item{\code{id}}{integer sample id number as referenced in the openai gpt2 sample file}
\item{\code{text}}{character - the text}
}}
\source{
\url{https://github.com/openai/gpt-2/tree/master/gpt-2-samples}
}
\usage{
unconditional_t07
}
\description{
Unconditionally generated texts from Open AI GPT-2
}
\keyword{datasets}
